suporte a decimais em raio, angulo e distancia

+++ projeto1/src/fr_tp1/IRobot.java
@@ -3,10 +3,10 @@ package fr_tp1;
 public interface IRobot {
     boolean ligar(String nome);
     void desligar();
-    void reta(int distancia);
-    void recuar(int distancia);
-    void curvarEsquerda(double raio, int angulo);
-    void curvarDireita(double raio, int angulo);
+    void reta(double distancia);
+    void recuar(double distancia);
+    void curvarEsquerda(double raio, double angulo);
+    void curvarDireita(double raio, double angulo);
     void parar(boolean travar);
     void setVelocidade(int vel);
     boolean isLigado();
